// src/@types/leaddropdown.d.ts
export type leaddropdownprops={
   
    errorText?:string| null,
    handledropdownclose:()=>void,
    leaddropdowns:boolean,
    leadid:number |null,
}
export type leaddataprops={
    leadStatusId:number,
    leadStatusName:string,

}
export type competitorprops={
    competitorId:number,
    competitorName:string,
}
export type enquiryprops={
    enquiryId:number,
    enquiryName:string,
}